Only light up around fireboxes
